This update adds the option of using HDF5's split virtual file driver (VFD). The split VFD in HDF5 allows one to write meta data to one file and raw data to another. The export of this HDF5 feature through silo involves a new driver type,
    DB_HDF5_SPLIT(MetaVFD,MetaInc,RawVFD,RawInc,ExtPair)

Where MetaVFD specifies the VFD for the meta-data stream, MetaInc
specifies the allocation incriment for the meta-data stream if
the MetaVFD specifies HDF5's core VFD, RawVFD/RawInc specify same
for raw data streams and ExtPair specifies the filename extension
pair (from amoung 16 possible) to use when naming the split files.

Several issues arise when using the split vfd due to the fact that
the string for filename passed to HDF5 to open is NOT necessarily
the name of any file on disk. That is because you pass the name
WITHOUT any special extensions specified by the split vfd. Browser
cannot correctly open a Silo file created with the split vfd without
also telling browser the filenaming extension conventions. The Silo
library specifies a default for these but allows applications to
define their own as well.

I also added a test for quad- and ucd-vars with more than 3 sub
variables.


git-svn-id: file:///usr/gapps/silo/svn/trunk@158 92378438-2262-0410-827d-edb41c1c42c8